Server/server/requirements.txt
Flask>=2.1
flask-cors>=3.0
numpy>=1.23
opencv-python>=4.7
face_recognition>=1.3.0
mysql-connector-python>=8.0
python-dotenv>=1.0
werkzeug>=2.1
Pillow>=9.0
# If you have a prebuilt dlib wheel (Windows), either install it manually or
# include the compatible dlib version here. A wheel is included in this repo
# (dlib-19.24.1-cp310-cp310-win_amd64.whl) — on Windows you can install it
# with: pip install dlib-19.24.1-cp310-cp310-win_amd64.whl
# Otherwise, list dlib here and ensure build tools (CMake + VS Build Tools)
# are available to compile it from source.
dlib==19.24.1

# Note: `face_recognition` depends on `dlib` which often requires
# compilation with Visual Studio Build Tools (Windows) or cmake + build-essentials (Linux).
# Installing dlib via wheel is easier if you have a compatible prebuilt wheel.